partial steps on way to throwing domino session exception all over

+++ src/main/java/mitll/langtest/server/database/userlist/IUserListDAO.java
@@ -86,7 +86,7 @@ public interface IUserListDAO extends IDAO {
 
   List<UserList<CommonShell>> getByName(long userid, String name, int projid);
 
-  boolean remove(long unique);
+  boolean remove(int unique);
 
   void bringBack(long unique);
